+++ plugins/pug.js
@@ -6,12 +6,16 @@ import { merge } from "../utils.js";
 // Default options
 const defaults = {
   extensions: [".pug"],
+  includes: null,
 };
 
 export default function (userOptions) {
-  const options = merge(defaults, userOptions);
-
   return (site) => {
+    const options = merge(
+      { ...defaults, includes: site.includes() },
+      userOptions,
+    );
+
     site.loadPages(options.extensions, loader, new Pug(site, pug));
   };
 }